requirements.txt
asgiref==3.6.0
certifi==2024.6.2
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.15
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
